# AggregatorStore to use ConcurrentDictionary