Swapped ConcurrentDictionary with regular dictionary with locks.
There Enumerator won't take "snapshot", so it'll be picking metric updates *after* enumeration started as well.

Before
|                    Method | WithSDK |     Mean |    Error |    StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |-------- |---------:|---------:|----------:|---------:|-------:|------:|------:|----------:|
| CounterWith3LabelsHotPath |    True | 726.1 ns | 67.75 ns | 185.47 ns | 655.2 ns |      - |     - |     - |         - |
| CounterWith5LabelsHotPath |    True | 831.2 ns | 20.00 ns |  58.34 ns | 817.9 ns | 0.0553 |     - |     - |     232 B

This PR 
|                    Method | WithSDK |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------------------- |-------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| CounterWith3LabelsHotPath |    True | 603.7 ns | 13.77 ns | 39.27 ns |      - |     - |     - |         - |
| CounterWith5LabelsHotPath |    True | 784.6 ns | 17.44 ns | 50.58 ns | 0.0553 |     - |     - |     232 B |

Based on Victor's experiment's earlier, ConcurrentDictionary performed poorly. The benchmarks above may not be the right way to validate this. Need to write stress tests where metrics updates are done from multiple thread.