@cijothomas 2021-09-17
closing in favor of https://github.com/open-telemetry/opentelemetry-dotnet/pull/2339